added front-end ui

+++ backend/controllers/goalController.js
@@ -1,10 +1,12 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 //@desc getgoals
 //@route GET /ap/goals
 //@access Pivate
 const getGoals = asyncHandler(async (req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
   
@@ -20,7 +22,8 @@ const setGoal = asyncHandler(async (req,res) => {
         throw new Error('please add a text field')
     }
     const goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user: req.user.id
     })
   
     res.status(201).json(goal);
@@ -35,6 +38,15 @@ const updateGoal = asyncHandler(async (req,res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+    if (goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorised')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -58,6 +70,16 @@ const deleteGoal = asyncHandler(async (req,res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+    if (goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorised')
     }
 
     await Goal.findByIdAndRemove(req.params.id, req.body, {
